requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
decorator==4.4.2
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
moviepy==1.0.3
numpy==1.25.0
opencv-python==4.7.0.72
Pillow==9.5.0
proglog==0.1.10
protobuf==4.21.12
protobuf-inspector==0.2
PyExecJS==1.5.1
requests==2.31.0
six==1.16.0
tqdm==4.65.0
urllib3==2.0.2
websocket-client==1.6.3
